moved execution under core
